+++ CHANGELOG.md
@@ -8,7 +8,37 @@ and this project adheres to
 
 ## [Unreleased]
 
-- 📝(doc) add publiccode.yml
+## Added
+
+- 📄(legal) Require contributors to sign a DCO #779
+
+## Changed
+
+- ♻️(frontend) Integrate UI kit #783
+
+## Fixed
+
+- 🐛(backend) compute ancestor_links in get_abilities if needed #725
+- 🔒️(back) restrict access to document accesses #801
+
+## [2.6.0] - 2025-03-21
+
+## Added
+
+- 📝(doc) add publiccode.yml #770
+
+## Changed
+
+- 🚸(frontend) ctrl+k modal not when editor is focused #712
+
+## Fixed
+
+- 🐛(back) allow only images to be used with the cors-proxy #781
+- 🐛(backend) stop returning inactive users on the list endpoint #636
+- 🔒️(backend) require at least 5 characters to search for users #636
+- 🔒️(back) throttle user list endpoint #636
+- 🔒️(back) remove pagination and limit to 5 for user list endpoint #636
+
 
 ## [2.5.0] - 2025-03-18
 
@@ -107,6 +137,8 @@ and this project adheres to
 
 ## Added
 
+- ✨(backend) add duplicate action to the document API endpoint
+- ⚗️(backend) add util to extract text from base64 yjs document
 - ✨(backend) add soft delete and restore API endpoints to documents #516
 - ✨(backend) allow organizing documents in a tree structure #516
 - ✨(backend) add "excerpt" field to document list serializer #516
@@ -470,7 +502,8 @@ and this project adheres to
 - ✨(frontend) Coming Soon page (#67)
 - 🚀 Impress, project to manage your documents easily and collaboratively.
 
-[unreleased]: https://github.com/numerique-gouv/impress/compare/v2.5.0...main
+[unreleased]: https://github.com/numerique-gouv/impress/compare/v2.6.0...main
+[v2.5.0]: https://github.com/numerique-gouv/impress/releases/v2.6.0
 [v2.5.0]: https://github.com/numerique-gouv/impress/releases/v2.5.0
 [v2.4.0]: https://github.com/numerique-gouv/impress/releases/v2.4.0
 [v2.3.0]: https://github.com/numerique-gouv/impress/releases/v2.3.0